archive/sagemaker/requirements-sagemaker.txt
numpy>=1.23.0,<2.0.0
cellpose>=2.0.0
Pillow>=9.0.0
opencv-python>=4.5.0
matplotlib>=3.5.0
imageio>=2.9.0
tqdm>=4.62.0
sagemaker>=2.0.0
boto3>=1.26.0
sagemaker-inference
requests>=2.28.0
